Assignment_4/requirements.txt
facenet_pytorch==2.6.0
matplotlib==3.8.4
pandas==2.2.2
Pillow==10.2.0
torch==2.2.0
tqdm==4.66.4
